Fix race condition in identifiers.SortIdentifiers()
The identifiers.SortIdentifiers() function is passed a list of strings
which it sorts internally, creating a race condition when this
function is called concurrently with the same list. (And it would be
called concurrently in practice, and is always passed the same list.)

The author of the function appears to have intended to make a copy of
the list to avoid this, but the "copy" was just a new slice pointing
into the old slice.